+++ HPXMLtoOpenStudio/resources/constants.rb
@@ -147,6 +147,10 @@ class Constants
     return "#{water_heater_name} dhw energy"
   end
 
+  def self.ObjectNameCombiSpaceHeatingEnergy(water_heater_name)
+    return "#{water_heater_name} htg energy"
+  end
+
   def self.ObjectNameComponentLoadsProgram
     return 'component loads program'
   end
@@ -248,11 +252,7 @@ class Constants
   end
 
   def self.ObjectNameIdealAirSystem
-    return 'ideal'
-  end
-
-  def self.ObjectNameIdealAirSystemResidual
-    return 'ideal residual'
+    return 'ideal air system'
   end
 
   def self.ObjectNameInfiltration
@@ -271,8 +271,12 @@ class Constants
     return 'mech vent'
   end
 
-  def self.ObjectNameMechanicalVentilationPreconditioning
-    return 'mech vent preconditioning'
+  def self.ObjectNameMechanicalVentilationPrecooling
+    return 'mech vent precooling'
+  end
+
+  def self.ObjectNameMechanicalVentilationPreheating
+    return 'mech vent preheating'
   end
 
   def self.ObjectNameMechanicalVentilationHouseFan